+++ src/EventStore.ClientAPI/WriteResult.cs
@@ -8,7 +8,7 @@ namespace EventStore.ClientAPI
         /// <summary>
         /// The next expected version for the stream.
         /// </summary>
-        public readonly int NextExpectedVersion;
+        public readonly long NextExpectedVersion;
 
         /// <summary>
         /// The <see cref="LogPosition"/> of the write.
@@ -20,7 +20,7 @@ namespace EventStore.ClientAPI
         /// </summary>
         /// <param name="nextExpectedVersion">The next expected version for the stream.</param>
         /// <param name="logPosition">The position of the write in the log</param>
-        public WriteResult(int nextExpectedVersion, Position logPosition)
+        public WriteResult(long nextExpectedVersion, Position logPosition)
         {
             LogPosition = logPosition;
             NextExpectedVersion = nextExpectedVersion;
